# api/openapi.yaml
openapi: 3.1.0
info:
  title: XLeaks API
  description: Decentralized P2P social platform API.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local node
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # Identity
    CreateIdentityRequest:
      type: object
      properties:
        passphrase:
          type: string
      required: [passphrase]
    CreateIdentityResponse:
      type: object
      properties:
        pubkey:
          type: string
          description: Hex-encoded public key
        address:
          type: string
        mnemonic:
          type: string
    ImportIdentityRequest:
      type: object
      properties:
        mnemonic:
          type: string
        seedPhrase:
          type: string
        passphrase:
          type: string
      required: [passphrase]
    ImportIdentityResponse:
      type: object
      properties:
        pubkey:
          type: string
        address:
          type: string
        status:
          type: string
    UnlockIdentityRequest:
      type: object
      properties:
        passphrase:
          type: string
      required: [passphrase]
    UnlockIdentityResponse:
      type: object
      properties:
        status:
          type: string
        pubkey:
          type: string
        address:
          type: string
...
    SwitchIdentityRequest:
      type: object
      properties:
        passphrase:
          type: string
      required: [passphrase]
    SwitchIdentityResponse:
      type: object
      properties:
        status:
          type: string
        pubkey:
          type: string
        address:
          type: string
    ExportIdentityResponse:
      type: object
      properties:
        pubkey:
          type: string
        address:
          type: string
        encrypted_key:
          type: string

    # Posts
...
  parameters:
...
    PubkeyParam:
      name: pubkey
      in: path
      required: true
      schema:
        type: string
      description: Hex-encoded public key
...
paths:
...
  /api/identity/create:
    post:
      tags: [Identity]
      summary: Create a new identity
      operationId: createIdentity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIdentityRequest'
      responses:
        '201':
          description: Identity created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateIdentityResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/identity/import:
    post:
      tags: [Identity]
      summary: Import an identity from mnemonic
      operationId: importIdentity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportIdentityRequest'
      responses:
        '200':
          description: Identity imported
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportIdentityResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/identity/unlock:
    post:
      tags: [Identity]
      summary: Unlock the active identity
      operationId: unlockIdentity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnlockIdentityRequest'
      responses:
        '200':
          description: Identity unlocked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UnlockIdentityResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Wrong passphrase
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/identity/list:
    get:
      tags: [Identity]
      summary: List all stored identities
      operationId: listIdentities
      responses:
        '200':
          description: List of identities
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/identity/switch/{pubkey}:
    put:
      tags: [Identity]
      summary: Switch to a different identity
      operationId: switchIdentity
      parameters:
        - $ref: '#/components/parameters/PubkeyParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SwitchIdentityRequest'
      responses:
        '200':
          description: Identity switched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SwitchIdentityResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Wrong passphrase
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/identity/export:
    post:
      tags: [Identity]
      summary: Export the active identity
      operationId: exportIdentity
      responses:
        '200':
          description: Exported identity file
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportIdentityResponse'
        '404':
          description: No active identity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Posts
...